Re-enable publish git status check

+++ publish.sh
@@ -1,10 +1,10 @@
 #!/bin/sh
 set -eu
 
-# if [[ $(git status -s) ]]; then
-# 	echo "The working directory is dirty. Please commit any pending changes."
-# 	exit 1
-# fi
+if [[ $(git status -s) ]]; then
+	echo "The working directory is dirty. Please commit any pending changes."
+	exit 1
+fi
 
 echo "Deleting old publication"
 rm -rf public
